+++ lib/tenstorrent/bh_arc/init.c
@@ -102,7 +102,7 @@ static void InitMrisc(void)
 {
 	static const char kMriscFwCfgTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "memfwcfg";
 	static const char kMriscFwTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "memfw";
-	uint32_t fw_size = 0;
+	size_t fw_size = 0;
 
 	for (uint8_t gddr_inst = 0; gddr_inst < 8; gddr_inst++) {
 		for (uint8_t noc2axi_port = 0; noc2axi_port < 3; noc2axi_port++) {
@@ -189,7 +189,7 @@ static void SerdesEthInit(void)
 
 	/* Load fw */
 	static const char kSerdesEthFwTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "ethsdfw";
-	uint32_t fw_size = 0;
+	size_t fw_size = 0;
 
 	if (tt_boot_fs_get_file(&boot_fs_data, kSerdesEthFwTag, large_sram_buffer, SCRATCHPAD_SIZE,
 				&fw_size) != TT_BOOT_FS_OK) {
@@ -211,7 +211,7 @@ static void EthInit(void)
 
 	/* Load fw */
 	static const char kEthFwTag[TT_BOOT_FS_IMAGE_TAG_SIZE] = "ethfw";
-	uint32_t fw_size = 0;
+	size_t fw_size = 0;
 
 	if (tt_boot_fs_get_file(&boot_fs_data, kEthFwTag, large_sram_buffer, SCRATCHPAD_SIZE,
 				&fw_size) != TT_BOOT_FS_OK) {
